requirements.txt
# Core
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1

# AI
google-generativeai==0.8.0

# RAG (optional — uncomment when needed)
# chromadb==0.5.0

# Frontend
streamlit==1.38.0

# Testing
pytest==8.3.0
